Add Ethereum prefix and hash with Keccak 256
